Ok, This commit represents 32 others. I deleted the previous repository because it exposed private information, but after managing to upload one single commit, I figured out it is going to take ages to upload every other commit. So there you go, 32 commits at once, from the 01/09/2020 to the 15/01/2021. Actually, I started working way before, but I was not using git. If you want to see a specific commit, just ask me, I still have them on my desk. Messages are (older first):
Made the logger a field of CustomLogger, not of bukkit.ScriptCommands, I haven't found a better place for it.

Create LICENSE

Lots of new things

Made big advancements to the ScriptValueList class and ScriptValueCollection.

Forgot to include a class, otherwise there are compile-time errors

Another huge commit
- debugged the ScriptValueList
- progress on Storage class
- limited variables in scripts to 100 characters
- fixed tests not being tested in mvn packages
- made a StringScriptValueMap (different from ScriptValueMap, but both are empty and to do)
- improved ScriptException
- I probably forgot other things, I should do smaller commits but more often lol

Debbuged some SQL things
SQLite-JDBC works better with the latest version
added a remove in ScriptValueList, that is todo
removed close method on statements that would cause ResultSet to break
still a lot to do for that ScriptValueMap and StringScriptValueMap

Finally a fully working ScriptValueList !

Made some method returing void as they modify the object itself

Added Copyright

Finished writing ScriptValueMap

Fixed makeSQL

Little bugs fixed

Finished writing StringScriptValueMap
Now, time for testing !

Finished ScriptValueMap and fixed ScriptThread

Fixed warnings and typos

tests and fixes with json

Made MySQl work
and changed the way global variables are loaded

Made a better logging system

A LOT of changes
- made MySQL
- reviewed the design of many classes, making Storage and Config static
- fixed lots of bugs

But I am still fixing that problem with every key of Maps in json being strings, which is not the case for ScriptValueMap.

little changes to the scripts' syntax

Bug fixes with json

things

Changed the way that arguments work

Finished Args and fixed some bugs
fixed Gson not saving a key when the value is null, and abandoned the idea of writing my own parser because even if there is json in json, it does not worth it to write a buggy parser. Also, ScriptValueCollections now don't clone everything they got added.

started events and made a big refactor

Put variables in Context and fixed contains method of ScriptValueCollections when using SQL

another bug fixed